GH-382: [refactor] Rename forgot password events package to follow naming conventions
