+++ Microservices/orchestration-service/src/main/java/app/sportahub/orchestrationservice/service/consumer/UserServiceConsumer.java
@@ -1,6 +1,6 @@
 package app.sportahub.orchestrationservice.service.consumer;
 
-import app.sportahub.kafka.events.ForgotPasswordRequestedEvent;
+import app.sportahub.kafka.events.forgotpassword.ForgotPasswordRequestedEvent;
 import app.sportahub.kafka.events.joinsporteventevent.JoinedEventsByUserRequestEvent;
 
 public interface UserServiceConsumer {
