+++ Microservices/orchestration-service/src/test/java/app/sportahub/orchestrationservice/kafka/consumer/UserServiceConsumerTest.java
@@ -1,8 +1,8 @@
 package app.sportahub.orchestrationservice.kafka.consumer;
 
 import app.sportahub.kafka.events.BaseEvent;
-import app.sportahub.kafka.events.ForgotPasswordRequestedEvent;
-import app.sportahub.kafka.events.ForgotPasswordSendEmailEvent;
+import app.sportahub.kafka.events.forgotpassword.ForgotPasswordRequestedEvent;
+import app.sportahub.kafka.events.forgotpassword.ForgotPasswordSendEmailEvent;
 import app.sportahub.orchestrationservice.service.consumer.UserServiceConsumerImpl;
 import app.sportahub.orchestrationservice.service.producer.EmailServiceProducer;
 import org.junit.jupiter.api.Assertions;
